BLD: meson: make version string and `__config__` generation work
