+++ setup.py
@@ -26,6 +26,9 @@ import warnings
 import sysconfig
 import importlib
 
+from tools.version_utils import (
+    get_version_info, write_version_py, IS_RELEASE_BRANCH
+)
 
 if sys.version_info[:2] < (3, 8):
     raise RuntimeError("Python version >= 3.8 required.")
@@ -54,52 +57,6 @@ Operating System :: MacOS
 
 """
 
-MAJOR = 1
-MINOR = 9
-MICRO = 0
-ISRELEASED = False
-IS_RELEASE_BRANCH = False
-VERSION = '%d.%d.%d' % (MAJOR, MINOR, MICRO)
-
-
-# Return the git revision as a string
-def git_version():
-    def _minimal_ext_cmd(cmd):
-        # construct minimal environment
-        env = {}
-        for k in ['SYSTEMROOT', 'PATH']:
-            v = os.environ.get(k)
-            if v is not None:
-                env[k] = v
-        # LANGUAGE is used on win32
-        env['LANGUAGE'] = 'C'
-        env['LANG'] = 'C'
-        env['LC_ALL'] = 'C'
-        out = subprocess.Popen(cmd, stdout=subprocess.PIPE, env=env).communicate()[0]
-        return out
-
-    try:
-        out = _minimal_ext_cmd(['git', 'rev-parse', 'HEAD'])
-        GIT_REVISION = out.strip().decode('ascii')[:7]
-
-        # We need a version number that's regularly incrementing for newer commits,
-        # so the sort order in a wheelhouse of nightly builds is correct (see
-        # https://github.com/MacPython/scipy-wheels/issues/114). It should also be
-        # a reproducible version number, so don't rely on date/time but base it on
-        # commit history. This gives the commit count since the previous branch
-        # point from the current branch (assuming a full `git clone`, it may be
-        # less if `--depth` was used - commonly the default in CI):
-        prev_version_tag = '^v{}.{}.0'.format(MAJOR, MINOR - 2)
-        out = _minimal_ext_cmd(['git', 'rev-list', 'HEAD', prev_version_tag,
-                                '--count'])
-        COMMIT_COUNT = out.strip().decode('ascii')
-        COMMIT_COUNT = '0' if not COMMIT_COUNT else COMMIT_COUNT
-    except OSError:
-        GIT_REVISION = "Unknown"
-        COMMIT_COUNT = "Unknown"
-
-    return GIT_REVISION, COMMIT_COUNT
-
 
 # BEFORE importing setuptools, remove MANIFEST. Otherwise it may not be
 # properly updated when the contents of directories change (true for distutils,
@@ -113,55 +70,6 @@ if os.path.exists('MANIFEST'):
 # a lot more robust than what was previously being used.
 builtins.__SCIPY_SETUP__ = True
 
-
-def get_version_info():
-    # Adding the git rev number needs to be done inside
-    # write_version_py(), otherwise the import of scipy.version messes
-    # up the build under Python 3.
-    FULLVERSION = VERSION
-    if os.path.exists('.git'):
-        GIT_REVISION, COMMIT_COUNT = git_version()
-    elif os.path.exists('scipy/version.py'):
-        # must be a source distribution, use existing version file
-        # load it as a separate module to not load scipy/__init__.py
-        import runpy
-        ns = runpy.run_path('scipy/version.py')
-        GIT_REVISION = ns['git_revision']
-        COMMIT_COUNT = ns['git_revision']
-    else:
-        GIT_REVISION = "Unknown"
-        COMMIT_COUNT = "Unknown"
-
-    if not ISRELEASED:
-        FULLVERSION += '.dev0+' + COMMIT_COUNT + '.' + GIT_REVISION
-
-    return FULLVERSION, GIT_REVISION, COMMIT_COUNT
-
-
-def write_version_py(filename='scipy/version.py'):
-    cnt = """
-# THIS FILE IS GENERATED FROM SCIPY SETUP.PY
-short_version = '%(version)s'
-version = '%(version)s'
-full_version = '%(full_version)s'
-git_revision = '%(git_revision)s'
-commit_count = '%(commit_count)s'
-release = %(isrelease)s
-
-if not release:
-    version = full_version
-"""
-    FULLVERSION, GIT_REVISION, COMMIT_COUNT = get_version_info()
-
-    a = open(filename, 'w')
-    try:
-        a.write(cnt % {'version': VERSION,
-                       'full_version': FULLVERSION,
-                       'git_revision': GIT_REVISION,
-                       'commit_count': COMMIT_COUNT,
-                       'isrelease': str(ISRELEASED)})
-    finally:
-        a.close()
 
 
 def check_submodules():
